==> ms_classification/02-data_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv("../data/df_requests.csv")
len(df)

df = df.drop_duplicates()
len(df)

import re

def remove_emails_com(text):
    text = str(text)
    return re.sub(r'[^\s]+.com', '', text).strip()

def remove_emails_br(text):
    text = str(text)
    return re.sub(r'[^\s]+.br', '', text).strip()

def remove_links(text):
    text = str(text)
    return re.sub(r'http[^\s]+', '', text).strip()

def remove_citations(text):
    text = str(text)
    return re.sub(r'@[^\s]+', '', text).strip()

def remove_numbers(text):
    text = str(text)
    return re.sub(r'\d+', '', text).strip()

def remove_special_characters(text):
    text = str(text)
    return re.sub(r'[^\w\s.!?]', '', text).strip()

def remove_empty_lines(text):
    text = str(text)
    return "\n".join([line for line in text.split('\n') if line.strip() != ''])

def remove_repeated_symbols(text):
    text = str(text)
    return re.sub(r'(\W)\1+', r'\1', text).strip()

def remove_excessive_spaces(text):
    text = str(text)
    return re.sub(r'\s+', ' ', text).strip()

def fix_isolated_commas(text):
    text = str(text)
    # Replace punctuation with a blank character before
    text = re.sub(r' ([.,:;!?])', r'\1', text)
    return text.strip()

from sklearn.pipeline import Pipeline 
from sklearn.preprocessing import FunctionTransformer


pipeline_clean_text = Pipeline([
    ('remove_links', FunctionTransformer(remove_links)),
    ('remove_emails_br', FunctionTransformer(remove_emails_br)),
    ('remove_emails_com', FunctionTransformer(remove_emails_com)),
    ('remove_citations', FunctionTransformer(remove_citations)),
    ('remove_excessive_spaces', FunctionTransformer(remove_excessive_spaces)),
    ('remove_repeated_symbols', FunctionTransformer(remove_repeated_symbols)),
    ('fix_isolated_commas', FunctionTransformer(fix_isolated_commas)),
    ('remove_numbers', FunctionTransformer(remove_numbers)),
    ('remove_special_characters', FunctionTransformer(remove_special_characters)),
    ('remove_empty_lines', FunctionTransformer(remove_empty_lines)),
])

df['DetalhamentoSolicitacao'][0], pipeline_clean_text.transform(df['DetalhamentoSolicitacao'][0])

df['DetalhamentoSolicitacao'] = df['DetalhamentoSolicitacao'].apply(pipeline_clean_text.transform)

!python -m spacy download pt_core_news_sm

import spacy

nlp = spacy.load('pt_core_news_sm')
stopwords_spacy = nlp.Defaults.stop_words
list(stopwords_spacy)[:10]

import nltk
from nltk.corpus import stopwords

nltk.download('stopwords')
stopwords_nltk = stopwords.words('portuguese')
list(stopwords_nltk)[:10]

both_stopwords = set(stopwords_nltk) | set(stopwords_spacy)

def remove_stop_word(text):
    text = re.sub(r'[^\w\s]', '', text)

    tokens = text.split()

    tokens = filter(lambda token: token not in both_stopwords, tokens)

    return ' '.join(tokens)

df['DetalhamentoSolicitacao'] = df['DetalhamentoSolicitacao'].str.lower().apply(remove_stop_word)
df['DetalhamentoSolicitacao']

def spacy_lemmatizer(text):
    doc = nlp(text)

    txt = [token.lemma_ for token in doc]

    txt = [word for word in txt if len(word) > 2]

    return ' '.join(txt)

df['DetalhamentoSolicitacao'] = df['DetalhamentoSolicitacao'].str.lower().apply(spacy_lemmatizer)

lengths = df['DetalhamentoSolicitacao'].dropna().str.len()

q1 = lengths.quantile(0.25)
q3 = lengths.quantile(0.75)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
iqr = q3 - q1
lower = q1 - 1.5 * iqr
upper = q3 + 1.5 * iqr
outliers = lengths[(lengths < lower) | (lengths > upper)]

indices = outliers.index.tolist()
df = df.drop(index=indices)
len(df)

df.dropna(subset=['DetalhamentoSolicitacao'], inplace=True)

df.to_csv("../data/df_requests_cleaned.csv")
