==> ms_classification/03-feature_selection.ipynb <==
import pandas as pd

df = pd.read_csv("../data/df_requests_cleaned.csv")
len(df)

X = df["DetalhamentoSolicitacao"]
len(X)

X.drop_duplicates()
X.dropna()
len(X)

df_new = df[['DetalhamentoSolicitacao', 'OrgaoDestinatario']].dropna().rename(
    columns={'DetalhamentoSolicitacao': 'x', 'OrgaoDestinatario': 'y'}
)
df_new.head()

df_new.to_csv("../data/df_requests_x_y.csv", index=False)
