==> ms_classification/06-inference_pipeline.ipynb <==
from joblib import load

vectorizer = load("../models/lai_ms_tfidf_vectorizer.joblib")
model, metadata = load("../models/lai_ms_model_classifier.joblib")

vectorizer, model, metadata

X_new = vectorizer.transform(["Gostaria acessar dados ministério"])
print("NNZ:", X_new.nnz)
print("Shape:", X_new.shape)

!python -m spacy download pt_core_news_sm

import spacy

nlp = spacy.load('pt_core_news_sm')
stopwords_spacy = nlp.Defaults.stop_words
list(stopwords_spacy)[:10]

import nltk
from nltk.corpus import stopwords

nltk.download('stopwords')
stopwords_nltk = stopwords.words('portuguese')
list(stopwords_nltk)[:10]

both_stopwords = set(stopwords_nltk) | set(stopwords_spacy)

import re

def remove_stop_word(text):
    text = re.sub(r'[^\w\s]', '', text)

    tokens = text.split()

    tokens = filter(lambda token: token not in both_stopwords, tokens)

    return ' '.join(tokens)

def spacy_lemmatizer(text):
    doc = nlp(text)

    txt = [token.lemma_ for token in doc]

    txt = [word for word in txt if len(word) > 2]

    return ' '.join(txt)

import pandas as pd

def predict(text, vectorizer, model):
    texts = pd.Series([text])
    no_stop_word_texts = texts.str.lower().apply(remove_stop_word)
    lemmatized_texts = no_stop_word_texts.str.lower().apply(spacy_lemmatizer)
    X_new = vectorizer.transform(lemmatized_texts)
    return X_new.nnz, model.predict_proba(X_new)

lais_outros = [
    "Eu tenho direito de acessar meus dados armazenados no sistema da justiça.",
    "Processo de número 500 me dá direito de acessar os dados gerados.",
    "Tenho meus direitos. Quero que vocês me enviem um relatório contendo os últimos processos judiciais no meu nome.",
    "solicitar acessar dados justiça",
    "solicitar acessar dados ministério justiça",
]

for lai in lais_outros:
    print(lai)
    nnz, pred_proba = predict(lai, vectorizer=vectorizer, model=model)
    print("NNZ:", nnz)
    print("Outro:", pred_proba[0][0])
    print("MS:", pred_proba[0][1])
    print()

lais_ms = [
    "Gostaria de solicitar meu prontuário médico, por gentiliza.",
    "Quero o prontuário gerado pelo médico Alberto Carlos que estava em plantão no dia 14. Ele me atendeu.",
    "Quero acesso aos medicamentos mais demandados pela polução no ano de 2024.",
    "Quero um relatório contendo as vacinas mais aplicadas no ano de 2024 em Santa Cruz, Rio Grande do Norte.",
    "Preciso de um relatório detalhado da frequências dos médicos em Unidades de Pronto Atendimento em Natal no último ano.",
]

for lai in lais_ms:
    print(lai)
    nnz, pred_proba = predict(lai, vectorizer=vectorizer, model=model)
    print("NNZ:", nnz)
    print("Outro:", pred_proba[0][0])
    print("MS:", pred_proba[0][1])
    print()

lai = input()
nnz, pred_proba = predict(lai, vectorizer=vectorizer, model=model)
print("NNZ:", nnz)
print("Outro:", pred_proba[0][0])
print("MS:", pred_proba[0][1])
print()
